refactor: extract useConflictDetection hook
